requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-can==4.2.2
requests==2.31.0
python-socketio==5.10.0
websockets==11.0.3